+++ TUNA/tuna_optfreq.py
@@ -2,8 +2,8 @@ import tuna_util as util
 import numpy as np
 import tuna_energy as energ
 import sys
-import tuna_postscf as postscf
 from termcolor import colored
+import tuna_postscf as postscf
 
 def print_trajectory(molecule, energy, coordinates):
 
@@ -23,19 +23,21 @@ def print_trajectory(molecule, energy, coordinates):
         file.close()
 
 
-def calculate_gradient(coordinates, calculation, atoms):
+def calculate_gradient(coordinates, calculation, atoms, silent=False):
 
     prod = 0.0001
 
     forward_coords = coordinates + np.array([[0, 0, 0], [0, 0, prod]])
     backward_coords = coordinates - np.array([[0, 0, 0], [0, 0, prod]])
 
-    scf_output_forward, _ = energ.calculate_energy(calculation, atoms, forward_coords, silent=True)
-    scf_output_backward, _ = energ.calculate_energy(calculation, atoms, backward_coords, silent=True)
+    if not silent: util.log(" Calculating energy on displaced geometry 1 of 2...  ", calculation, end=""); sys.stdout.flush()
+    _, _, forward_energy, _ = energ.calculate_energy(calculation, atoms, forward_coords, silent=True)
+    if not silent: util.log("[Done]", calculation)
 
-    forward_energy = scf_output_forward.energy
-    backward_energy = scf_output_backward.energy
-
+    if not silent: util.log(" Calculating energy on displaced geometry 2 of 2...  ", calculation, end=""); sys.stdout.flush()
+    _, _, backward_energy, _ = energ.calculate_energy(calculation, atoms, backward_coords, silent=True)
+    if not silent: util.log("[Done]", calculation)
+    
     gradient = (forward_energy - backward_energy) / (2 * prod)
 
     return gradient
@@ -50,24 +52,36 @@ def calculate_approximate_hessian(delta_x, delta_grad):
 
 
 
-def calculate_hessian(energy, coordinates, calculation, atoms):
-    
+def calculate_hessian(coordinates, calculation, atoms, silent=False):
+
     prod = 0.0001
+    prodding_coords = np.array([[0,0,0],[0,0, prod]])  
 
-    far_forward_coords = coordinates + np.array([[0,0,0],[0,0, 2 * prod]])  
-    forward_coords = coordinates + np.array([[0,0,0],[0,0, prod]])  
-    backward_coords = coordinates - np.array([[0,0,0],[0,0, prod]])   
-    far_backward_coords = coordinates - np.array([[0,0,0],[0,0, 2 * prod]])    
 
-    scf_output_far_forward, _  = energ.calculate_energy(calculation, atoms, far_forward_coords, silent=True)   
-    scf_output_forward, _  = energ.calculate_energy(calculation, atoms, forward_coords, silent=True)
-    scf_output_backward, _  = energ.calculate_energy(calculation, atoms, backward_coords, silent=True)
-    scf_output_far_backward, _  = energ.calculate_energy(calculation, atoms, far_backward_coords, silent=True)
+    far_forward_coords = coordinates + 2 * prodding_coords
+    forward_coords = coordinates + prodding_coords
+    backward_coords = coordinates - prodding_coords
+    far_backward_coords = coordinates - 2 * prodding_coords  
 
-    far_forward_energy = scf_output_far_forward.energy
-    forward_energy = scf_output_forward.energy
-    backward_energy = scf_output_backward.energy
-    far_backward_energy = scf_output_far_backward.energy
+    if not silent: util.log("\n Calculating energy on displaced geometry 1 of 5...  ", calculation, end=""); sys.stdout.flush()
+    _, _, far_forward_energy, _ = energ.calculate_energy(calculation, atoms, far_forward_coords, silent=True)
+    if not silent: util.log("[Done]", calculation)   
+
+    if not silent: util.log(" Calculating energy on displaced geometry 2 of 5...  ", calculation, end=""); sys.stdout.flush()
+    _, _, forward_energy, _ = energ.calculate_energy(calculation, atoms, forward_coords, silent=True)
+    if not silent: util.log("[Done]", calculation)   
+
+    if not silent: util.log(" Calculating energy on displaced geometry 3 of 5...  ", calculation, end=""); sys.stdout.flush()
+    _, _, energy, _ = energ.calculate_energy(calculation, atoms, coordinates, silent=True)
+    if not silent: util.log("[Done]", calculation)   
+
+    if not silent: util.log(" Calculating energy on displaced geometry 4 of 5...  ", calculation, end=""); sys.stdout.flush()
+    _, _, backward_energy, _ = energ.calculate_energy(calculation, atoms, backward_coords, silent=True)
+    if not silent: util.log("[Done]", calculation)   
+
+    if not silent: util.log(" Calculating energy on displaced geometry 5 of 5...  ", calculation, end=""); sys.stdout.flush()
+    _, _, far_backward_energy, _  = energ.calculate_energy(calculation, atoms, far_backward_coords, silent=True)
+    if not silent: util.log("[Done]\n", calculation)   
 
     hessian = (-far_forward_energy + 16 * forward_energy - 30 * energy + 16 * backward_energy - far_backward_energy) / (12 * prod ** 2)
 
@@ -100,34 +114,35 @@ def optimise_geometry(calculation, atoms, starting_coordinates):
     print(f"Maximum iterations: {max_geom_iter}")
     print(f"Maximum step: {util.bohr_to_angstrom(maximum_step):.5f}")
 
-    P_guess = None; E_guess = 0
+    P_guess = None; E_guess = 0; P_guess_alpha = None; P_guess_beta = None
 
     for iteration in range(1, max_geom_iter + 1):
 
         print(f"\n~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~")
         print(f"Beginning energy and gradient calculation on geometry iteration number {iteration}...")
-        print(f"~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~\n")
+        print(f"~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~")
         
         if not calculation.moread: 
 
             P_guess = None
+            P_guess_alpha = None
+            P_guess_beta = None
             E_guess = 0
 
-        if calculation.additional_print: scf_output, molecule = energ.calculate_energy(calculation, atoms, coordinates, P_guess, E_guess, terse=False)
-        else: scf_output, molecule = energ.calculate_energy(calculation, atoms, coordinates, P_guess, E_guess, terse=True)
+        if calculation.additional_print: scf_output, molecule, energy, final_P = energ.calculate_energy(calculation, atoms, coordinates, P_guess, P_guess_alpha=P_guess_alpha, P_guess_beta=P_guess_beta, E_guess=E_guess, terse=False)
+        else: scf_output, molecule, energy, final_P = energ.calculate_energy(calculation, atoms, coordinates, P_guess, P_guess_alpha=P_guess_alpha, P_guess_beta=P_guess_beta, E_guess=E_guess, terse=True)
 
+        P = final_P
+        P_guess =  scf_output.P
+        P_guess_alpha = scf_output.P_alpha
+        P_guess_beta = scf_output.P_beta
 
-        energy = scf_output.energy
-        P = scf_output.P
-
-        P_guess = P
         E_guess = energy
 
 
-        print("\n Calculating numerical gradient...  ", end=""); sys.stdout.flush()
+        print("\n Beginning numerical gradient calculation...  \n")
         gradient = calculate_gradient(coordinates, calculation, atoms)
 
-        print("[Done]")
 
         bond_length = molecule.bond_length
 
@@ -138,7 +153,9 @@ def optimise_geometry(calculation, atoms, starting_coordinates):
      
         if iteration > 1:
 
-            if calculation.calchess: h = calculate_hessian(energy, coordinates, calculation, atoms)
+            if calculation.calchess: 
+                util.log("\n Beginning calculation of exact Hessian...", calculation)
+                h = calculate_hessian(coordinates, calculation, atoms)
             else: h = calculate_approximate_hessian(bond_length - old_bond_length, gradient - old_gradient)
 
 
@@ -174,12 +191,12 @@ def optimise_geometry(calculation, atoms, starting_coordinates):
         if converged_grad and converged_step: 
 
             print("\n==========================================")           
-            print(f" Optimisation converged in {iteration} iterations!")
+            print(colored(f" Optimisation converged in {iteration} iterations!","white"))
             print("==========================================")
 
-            postscf.post_scf_output(molecule, calculation, scf_output.epsilons, scf_output.molecular_orbitals, P, scf_output.S, scf_output.ao_ranges, scf_output.D, scf_output.P_alpha, scf_output.P_beta)
+            postscf.post_scf_output(molecule, calculation, scf_output.epsilons, scf_output.molecular_orbitals, P, scf_output.S, molecule.ao_ranges, scf_output.D, scf_output.P_alpha, scf_output.P_beta)
             
-            print(f"\n Optimisation converged in {iteration} iterations to bond length of {util.bohr_to_angstrom(bond_length):.4f} Angstroms!")
+            print(f"\n Optimisation converged in {iteration} iterations to bond length of {util.bohr_to_angstrom(bond_length):.6f} angstroms!")
             print(f"\n Final single point energy: {energy:.10f}")
 
             return molecule, energy
@@ -215,8 +232,7 @@ def calculate_frequency(calculation, atoms=None, coordinates=None, optimised_mol
 
     if calculation.calculation_type == "FREQ":
           
-        scf_output, molecule = energ.calculate_energy(calculation, atoms, coordinates)
-        energy = scf_output.energy
+        _, molecule, energy, _ = energ.calculate_energy(calculation, atoms, coordinates)
     
     else:
 
@@ -239,13 +255,10 @@ def calculate_frequency(calculation, atoms=None, coordinates=None, optimised_mol
     print("~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~")
 
     
-    print(f"\n Calculating numerical Hessian at bond length of {util.bohr_to_angstrom(bond_length):.4f} Angstroms...  ", end=""); sys.stdout.flush()
+    print(f"\n Hessian will be calculated at a bond length of {util.bohr_to_angstrom(bond_length):.6f} angstroms.")
     
-    k = calculate_hessian(energy, coordinates, calculation, atoms)
-    
-    print("[Done]\n")
-    
-    
+    k = calculate_hessian(coordinates, calculation, atoms)
+
     reduced_mass = postscf.calculate_reduced_mass(masses)
 
     if k > 0:
@@ -262,7 +275,7 @@ def calculate_frequency(calculation, atoms=None, coordinates=None, optimised_mol
         vibrational_entropy = 0; 
         vibrational_internal_energy = 0
         
-        util.warning("Imaginary frequency calculated! Zero-point energy and vibrational thermochemistry set to zero!")
+        util.warning("Imaginary frequency calculated! Zero-point energy and vibrational thermochemistry set to zero!\n")
 
 
     frequency_per_cm = frequency_hartree * util.constants.per_cm_in_hartree
@@ -287,7 +300,7 @@ def calculate_frequency(calculation, atoms=None, coordinates=None, optimised_mol
     print(" Entropies multiplied by temperature to give units of energy.")
     print(f" Using symmetry number derived from {point_group} point group for rotational entropy.")
 
-    rotational_constant_per_cm, rotational_constant_GHz = postscf.calculate_rotational_constant(masses, coordinates)
+    rotational_constant_per_cm, _ = postscf.calculate_rotational_constant(masses, coordinates)
 
     U, translational_internal_energy, rotational_internal_energy, vibrational_internal_energy = thermo.calculate_internal_energy(energy, zpe, temp, frequency_per_cm)
     H = thermo.calculate_enthalpy(U, temp)
@@ -311,5 +324,5 @@ def calculate_frequency(calculation, atoms=None, coordinates=None, optimised_mol
     print(f"  Vibrational energy:    {vibrational_internal_energy:.10f}       Vibrational entropy:     {temp*vibrational_entropy:.10f}  ")
     print(f"\n  Internal energy:    {space_1}  {U:.10f}")
     print(f"  Enthalpy:         {space_2}    {H:.10f}       Entropy:                 {temp*S:.10f}")
-    print(f"\n  Gibbs free energy:    {G:.10f}      Non-electronic energy:   {energy - G:.10f}")
+    print(f"\n  Gibbs free energy:    {G:.10f}       Non-electronic energy:   {energy - G:.10f}")
     print(" ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~")